+++ compile/compile.bat
@@ -1,6 +1,6 @@
 git pull
-pyinstaller -F --add-data includes\undelivered_help.pdf;includes --distpath bin/executables npbc.py
+pyinstaller -F --add-data includes\undelivered_help.pdf;includes -nnpbc --distpath bin/executables npbc.py
 git add .
 git commit -m "windows compile"
 git pull
-git push origin master
+git push origin
