+++ compile/compile.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 git pull
-pyinstaller -F --add-data includes/undelivered_help.pdf:includes --distpath bin/executables npbc.py
+pyinstaller -F --add-data includes/undelivered_help.pdf:includes -n linnpbc --distpath bin/executables npbc.py
 git add .
 git commit -m "linux on x64 compile"
 git pull
-git push origin master
+git push origin
