+++ install/install.sh
@@ -4,6 +4,6 @@ dir=~/.npbc
 if [[ ! -d "$dir" ]]; then
     cp -r ~/bin/npbc/data ~/.npbc/
 fi
-chmod +x ~/bin/npbc/bin/npbc.sh
-echo 'alias npbc=~/bin/npbc/bin/npbc.sh' >> ~/.bashrc
-alias npbc=~/bin/npbc/bin/npbc.sh
+chmod +x ~/bin/npbc/bin/linnpbc.sh
+echo 'alias npbc=~/bin/npbc/bin/linnpbc.sh' >> ~/.bashrc
+alias npbc=~/bin/npbc/bin/linnpbc.sh
